refactor(ios): extract message subviews from MobileChatView (#14264)
## Summary

Iteration 7 of the iOS refactor loop (no Linear issue — ad-hoc loop work). Behavior-preserving split of `Features/Chat/MobileChatView.swift` (**700 → 221 lines**).

- The message rendering stack moves to a new `Features/Chat/MobileChatMessageViews.swift` (480 lines): `MobileChatMessageRow`, `MobileChatThinkingDotsView`, `ChatComposerView`, `MobileChatProseTone`, `MobileChatFlowToken`, the custom `MobileChatInlineFlowLayout` `Layout`, `MobileChatEntityChipView`, and `MobileChatProseText`.
- `MobileChatView.swift` keeps the keyboard policy + the top-level view (timeline scroll, composer wiring).
- Access changes: only `MobileChatMessageRow` and `ChatComposerView` go `private` → internal (referenced from `MobileChatView`); all other bodies verbatim. Dropped the now-unused `UIKit` import from `MobileChatView.swift` (pasteboard usage moved with the subviews).
- `project.pbxproj` wired with explicit entries (`plutil -lint` clean).

## Verification

- `pnpm run ios:lint` → **clean ✓**
- `pnpm run ios:test` (iPhone 17 sim, iOS 26.3.1) → **248 passed, 6 skipped, 2 failed** — the known local-sim-only pair tracked in **JOV-4252** (passes on GitHub CI).
- `MobileChatContentParserTests` (exercises `MobileChatProseText.flowTokens`) passes.

## Risks

- Low: same-module move; two access-level widenings, no logic changes.
